requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
google-generativeai==0.3.0
python-pptx==0.6.21
moviepy==1.0.3
Pillow==9.0.0
requests==2.26.0
gTTS==2.2.3
python-slugify==8.0.4
pytest==8.0.2
black==24.2.0
flake8==7.0.0
ffmpeg-python==0.2.0 